neuro_calc/pyproject.toml
[tool.poetry]
name = "neuro_calc"
version = "0.1.0"
package-mode = false
description = "Invariant Spatio-Temporal Hand Calculation Engine"
authors = ["Antoine Besson antoine.besson@telecom-paris.fr"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"  # Lock Python version to ensure Torch compatibility
numpy = "^1.24"
opencv-python = "^4.8"
mediapipe = "^0.10.14"
hydra-core = "^1.3"
# PyTorch Geometric stack (install via specialized pip command usually, but listed here for tracking)
torch = "^2.0"
protobuf = "^4.25"
tqdm = "^4.67.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
black = "^23.0"        # Auto-formatter
isort = "^5.12"        # Import sorter

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"